Use new Inflector constructor in run context and add integration test case for the bug
`payment_details` in `custom_inflections.yml` is an "uncountable",
which means it should not be inflected when found in a constant association.

Without the fix for this bug, this check run would pass because
Packwerk is unable to resolve the constant and treats it as a
non-violation. In reality, `payment_details` exists as a constant,
which makes it a legitimate violation. See:

https://github.com/Shopify/packwerk/issues/48