Remove test that was testing the behaviour of the bug
Previously `custom_inflections.yml` was not specified in the
skeleton configuration file that is used in the integration tests.

In this test, the file was added to the configuration
and expected to fail. This was because the validator was using
`ActiveSupport::Inflector` and not checking for any custom inflections.

This test is already covered in the `ApplicationValidation` checker,
`#check_inflections_file`.